requirements.txt
fastapi
uvicorn
lark-oapi==1.4.18
pydantic
pydantic-settings
ruamel.yaml
httpx
python-dotenv
flask
apscheduler
